+++ spec/Plugin/RetryPluginSpec.php
@@ -30,7 +30,9 @@ class RetryPluginSpec extends ObjectBehavior
             }
         };
 
-        $this->handleRequest($request, $next, function () {})->shouldReturnAnInstanceOf('Http\Client\Promise\HttpFulfilledPromise');
+        $promise = $this->handleRequest($request, $next, function () {});
+        $promise->shouldReturnAnInstanceOf('Http\Client\Common\Deferred');
+        $promise->wait()->shouldReturn($response);
     }
 
     function it_throws_exception_on_multiple_exceptions(RequestInterface $request)
@@ -53,7 +55,7 @@ class RetryPluginSpec extends ObjectBehavior
         };
 
         $promise = $this->handleRequest($request, $next, function () {});
-        $promise->shouldReturnAnInstanceOf('Http\Client\Promise\HttpRejectedPromise');
+        $promise->shouldReturnAnInstanceOf('Http\Client\Common\Deferred');
         $promise->shouldThrow($exception2)->duringWait();
     }
 
@@ -76,7 +78,7 @@ class RetryPluginSpec extends ObjectBehavior
         };
 
         $promise = $this->handleRequest($request, $next, function () {});
-        $promise->shouldReturnAnInstanceOf('Http\Client\Promise\HttpFulfilledPromise');
+        $promise->shouldReturnAnInstanceOf('Http\Client\Common\Deferred');
         $promise->wait()->shouldReturn($response);
     }
 
@@ -98,8 +100,13 @@ class RetryPluginSpec extends ObjectBehavior
             }
         };
 
-        $this->handleRequest($request, $next, function () {})->shouldReturnAnInstanceOf('Http\Client\Promise\HttpFulfilledPromise');
-        $this->handleRequest($request, $next, function () {})->shouldReturnAnInstanceOf('Http\Client\Promise\HttpFulfilledPromise');
+        $promise = $this->handleRequest($request, $next, function () {});
+        $promise->shouldReturnAnInstanceOf('Http\Client\Common\Deferred');
+        $promise->wait()->shouldReturn($response);
+
+        $promise = $this->handleRequest($request, $next, function () {});
+        $promise->shouldReturnAnInstanceOf('Http\Client\Common\Deferred');
+        $promise->wait()->shouldReturn($response);
     }
 
     function it_has_an_exponential_default_delay(RequestInterface $request, Exception\HttpException $exception)